Add plotting functions
Plotting meshes is important for our current registration pipeline. I am moving
library functions from mri-registration to this repo, as this repo is intended
to be a library of functions for dealing with MRI data and performing
registrations of MRI data.

This adds `plot_tri_mesh`, as well as some utility functions
`make_3d_ax_look_normal` and `set_axes_equal` to this repo.