Add reticle calibration and rotations (#28)
Add code from the mri-registration rotation file, and also code for fitting rotations to reticle calibrations.

In the process, I cleaned up both sets of code. In particular, the outputs of fit_rotation_params have been changed to return the regular rotation matrix and the translation in the probe frame. The old outputs are accessible by using the legacy_outputs=True parameter.